Move to parcel for compileing js
