+++ package.json
@@ -2,9 +2,11 @@
     "name": "rhgservices",
     "version": "1.0.0",
     "description": "",
+    "browserslist": "> 0.5%, last 2 versions, not dead",
     "scripts": {
         "build": "echo \"can't build yet\" && exit 1"
     },
+    "type": "module",
     "repository": {
         "type": "git",
         "url": "git+https://github.com/bengosney/rhgservices.git"
@@ -24,10 +26,12 @@
         "gulp-rename": "^2.0.0",
         "gulp-sass": "^5.1.0",
         "gulp-sourcemaps": "^3.0.0",
+        "parcel": "^2.8.3",
         "postcss": "^8.4.12",
         "postcss-combine-media-query": "^1.0.1",
         "postcss-preset-env": "^7.4.2",
         "sass": "^1.49.9",
+        "tslib": "^2.5.0",
         "typescript": "^4.9.5"
     }
 }
